% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroweight_uitility.R
\name{process_input}
\alias{process_input}
\title{Process Input GIS files into target format and clip}
\usage{
process_input(
  input = NULL,
  input_name = NULL,
  variable_names = NULL,
  target = NULL,
  clip_region = NULL,
  resample_type = c("bilinear", "near"),
  ...
)
}
\description{
Process Input GIS files into target format and clip
}
