% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_input.R
\name{process_input}
\alias{process_input}
\title{Process Input GIS files into target format and clip}
\usage{
process_input(
  input = NULL,
  input_name = NULL,
  variable_names = NULL,
  target = NULL,
  clip_region = NULL,
  resample_type = c("bilinear", "near"),
  working_dir = NULL,
  ...
)
}
\arguments{
\item{input}{character. (full file path with extension, e.g., "C:/Users/Administrator/Desktop/input.shp"), \code{sf}, \code{SpatVector}, \code{PackedSpatVector}, \code{RasterLayer}, \code{SpatRaster}, or \code{PackedSpatRaster}. Input GIS data.}

\item{input_name}{\code{NULL} or character. A name that is used only in error or warning messages.}

\item{variable_names}{\code{NULL} or character. The names if input that should be included in the coerced output. If \code{NULL} all variables in \code{input} are used.}

\item{target}{\code{NULL} or character (full file path with extension, e.g., "C:/Users/Administrator/Desktop/target.tif"), \code{sf}, \code{SpatVector}, \code{PackedSpatVector}, \code{RasterLayer}, \code{SpatRaster}, or \code{PackedSpatRaster}. Target for intput to be coerced into. If \code{NULL} all variables in \code{input} are returned as \code{SpatVector} or \code{SpatRaster}.}

\item{clip_region}{\code{NULL} or character (full file path with extension, e.g., "C:/Users/Administrator/Desktop/clip_region.shp"), \code{sf}, \code{SpatVector}, \code{PackedSpatVector}, \code{RasterLayer}, \code{SpatRaster}, or \code{PackedSpatRaster}. Input with be clipped and masked to this region. Internally converted to  \code{SpatVector}. If \code{NULL}, full extent of \code{input} is returned.}

\item{resample_type}{character. If \code{bilinear} (default), the \code{variable_names} being coerced are numeric, if \code{'near'}, the \code{variable_names} being coerced are categorical, for the purposes of resampling/projecting if \code{target} is \code{SpatRaster}.}

\item{working_dir}{\code{NULL} or character. Folder path for temporary file storage. If \code{NULL}, assigned internally.}

\item{...}{other variables passed to various \code{terra} functions.}
}
\value{
an object of class \code{SpatVector} or \code{SpatRaster} depending on \code{target}
}
\description{
\code{hydroweight::process_input()} processes any input GIS data files into type \code{SpatVector} or \code{SpatRaster} depending on \code{target}.
}
\details{
\code{input} data is coerced to type \code{SpatVector} or \code{SpatRaster} depending on class of \code{target}.
If \code{resample_type} == \code{bilinear}, and \code{target} is of class \code{SpatRaster}, \code{input} is
coerced to class \code{SpatRaster}, reasterizing and/or reprojecting as necessary. If \code{resample_type} == \code{near},
\code{target} is of class \code{SpatRaster}, and \code{input} is of class \code{SpatRaster}, unique values of \code{input} are coerced
into separate layers of a \code{SpatRaster}, with 1 indicating the value is present; if input\verb{is of class \code{SpatVector}, unique values of}variable_names` are each coerced into separate layers in a \code{SpatRaster}, with 1 indicating the value is present.
}
